+++ tests/plugins/manifest_fixtures.py
@@ -1,11 +1,8 @@
 """Plugin module for mocking S3 buckets"""
-# Standard
 import os
 
-# Installed
 import pytest
 
-# Local
 from libera_utils.aws.constants import ManifestType
 from libera_utils.io.manifest import Manifest