+++ tests/unit/test_quality_flags.py
@@ -1,9 +1,9 @@
 """Test coverage for the libera_utils.quality_flags module"""
-# Installed
+from enum import NAMED_FLAGS, UNIQUE, verify
+
 import pytest
-# Local
+
 import libera_utils.quality_flags as qf
-from enum import verify, NAMED_FLAGS, UNIQUE
 
 
 def test_quality_flag_verify_unique():
@@ -11,7 +11,7 @@ def test_quality_flag_verify_unique():
     Test to ensure that all bit masks are named uniquely
     """
 
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match="aliases found"):
         @verify(UNIQUE)
         class TestFlag(qf.LiberaFlag):
             BIT_0 = 0b001
@@ -20,12 +20,13 @@ def test_quality_flag_verify_unique():
             BIT_1_AND_2 = 0b110
             ALIAS = 0b100  # NOT unique
 
+
 def test_quality_flag_verify_named_flags():
     """
     Test to ensure that all individual bits are named in the enum before being referenced in
     combined flag bit masks
     """
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match="invalid Flag 'TestFlag': alias BIT_1_AND_2 is missing value 0x4"):
         @verify(NAMED_FLAGS)
         class TestFlag(qf.LiberaFlag):
             BIT_0 = 0b001
@@ -42,7 +43,7 @@ def test_quality_flag_member_immutability():
         BIT_2 = 0b100
         BIT_1_AND_2 = 0b110
 
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match="invalid value 99"):
         TestFlag(99)  # Raises because 99 is out of range (violates STRICT boundary)
 
     with pytest.raises(AttributeError):
@@ -96,10 +97,10 @@ def test_strict_quality_flags():
     TestFlag(1)  # bit 0
     TestFlag(4)  # bit 2
     TestFlag(5)  # bit 0 and 2
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match="invalid value"):
         TestFlag(2)
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match="invalid value"):
         TestFlag(3)
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match="invalid value"):
         TestFlag(6)